refactor: extract user profile fields
