+++ src/app/Scenes/MyProfile/MyProfileEditForm.tsx
@@ -6,7 +6,6 @@ import {
   CheckCircleFillIcon,
   CheckCircleIcon,
   Flex,
-  Input,
   Join,
   Message,
   Spacer,
@@ -16,14 +15,18 @@ import {
 } from "@artsy/palette-mobile"
 import { useActionSheet } from "@expo/react-native-action-sheet"
 import { useNavigation } from "@react-navigation/native"
-import { EditableLocation } from "__generated__/ConfirmBidUpdateUserMutation.graphql"
 import { MyProfileEditFormQuery } from "__generated__/MyProfileEditFormQuery.graphql"
 import { MyProfileEditForm_me$key } from "__generated__/MyProfileEditForm_me.graphql"
 import { Image } from "app/Components/Bidding/Elements/Image"
 import { FancyModalHeader } from "app/Components/FancyModal/FancyModalHeader"
-import { LocationAutocomplete, buildLocationDisplay } from "app/Components/LocationAutocomplete"
+import { buildLocationDisplay } from "app/Components/LocationAutocomplete"
 import LoadingModal from "app/Components/Modals/LoadingModal"
 import { updateMyUserProfile } from "app/Scenes/MyAccount/updateMyUserProfile"
+import {
+  UserProfileFields,
+  UserProfileFormikSchema,
+  userProfileYupSchema,
+} from "app/Scenes/MyProfile/Components/UserProfileFields"
 import { navigate } from "app/system/navigation/navigate"
 import { ArtsyKeyboardAvoidingView } from "app/utils/ArtsyKeyboardAvoidingView"
 import { storeLocalImage, useLocalImageStorage } from "app/utils/LocalImageStore"
@@ -32,8 +35,8 @@ import { PlaceholderBox, PlaceholderText, ProvidePlaceholderContext } from "app/
 import { showPhotoActionSheet } from "app/utils/requestPhotos"
 import { sendEmail } from "app/utils/sendEmail"
 import { useHasBeenTrue } from "app/utils/useHasBeenTrue"
-import { useFormik } from "formik"
-import React, { Suspense, useEffect, useRef, useState } from "react"
+import { FormikProvider, useFormik } from "formik"
+import React, { Suspense, useEffect, useState } from "react"
 import { InteractionManager, ScrollView } from "react-native"
 import { graphql, useLazyLoadQuery, useRefetchableFragment } from "react-relay"
 import { useTracking } from "react-tracking"
@@ -42,21 +45,12 @@ import { useHandleEmailVerification, useHandleIDVerification } from "./useHandle
 
 const ICON_SIZE = 22
 
-interface EditableLocationProps extends EditableLocation {
-  display: string | null
-}
-interface EditMyProfileValuesSchema {
+interface EditMyProfileValuesSchema extends UserProfileFormikSchema {
   photo: string
-  name: string
-  displayLocation: { display: string | null }
-  location: Partial<EditableLocationProps> | null | undefined
-  profession: string
-  otherRelevantPositions: string
 }
 
-const editMyProfileSchema = Yup.object().shape({
+const editMyProfileSchema = userProfileYupSchema.shape({
   photo: Yup.string(),
-  name: Yup.string().required("Name is required"),
 })
 
 interface MyProfileEditFormProps {
@@ -76,11 +70,6 @@ export const MyProfileEditForm: React.FC<MyProfileEditFormProps> = ({ onSuccess 
   const navigation = useNavigation()
 
   const { showActionSheetWithOptions } = useActionSheet()
-
-  const nameInputRef = useRef<Input>(null)
-  const relevantPositionsInputRef = useRef<Input>(null)
-  const professionInputRef = useRef<Input>(null)
-  const locationInputRef = useRef<Input>(null)
 
   const [refreshKey, setRefreshKey] = useState(0)
   const [loading, setLoading] = useState<boolean>(false)
@@ -130,45 +119,46 @@ export const MyProfileEditForm: React.FC<MyProfileEditFormProps> = ({ onSuccess 
     }
   }
 
-  const { handleSubmit, handleChange, setFieldValue, dirty, values, errors, validateForm } =
-    useFormik<EditMyProfileValuesSchema>({
-      enableReinitialize: true,
-      validateOnChange: true,
-      validateOnBlur: true,
-      initialValues: {
-        name: me?.name ?? "",
-        displayLocation: { display: buildLocationDisplay(me?.location ?? null) },
-        location:
-          {
-            ...me?.location,
-          } ?? undefined,
-        profession: me?.profession ?? "",
-        otherRelevantPositions: me?.otherRelevantPositions ?? "",
-        photo: localImage?.path || me?.icon?.url || "",
-      },
-      initialErrors: {},
-      onSubmit: async ({ photo, ...otherValues }) => {
-        try {
-          setLoading(true)
-          await Promise.all([
-            updateUserInfo(otherValues),
-            didUpdatePhoto && uploadProfilePhoto(photo),
-          ])
+  const formikBag = useFormik<EditMyProfileValuesSchema>({
+    enableReinitialize: true,
+    validateOnChange: true,
+    validateOnBlur: true,
+    initialValues: {
+      name: me?.name ?? "",
+      displayLocation: { display: buildLocationDisplay(me?.location ?? null) },
+      location:
+        {
+          ...me?.location,
+        } ?? undefined,
+      profession: me?.profession ?? "",
+      otherRelevantPositions: me?.otherRelevantPositions ?? "",
+      photo: localImage?.path || me?.icon?.url || "",
+    },
+    initialErrors: {},
+    onSubmit: async ({ photo, ...otherValues }) => {
+      try {
+        setLoading(true)
+        await Promise.all([
+          updateUserInfo(otherValues),
+          didUpdatePhoto && uploadProfilePhoto(photo),
+        ])
 
-          trackEvent(tracks.editedUserProfile())
-        } catch (error) {
-          console.error("Failed to update profile", error)
-        } finally {
-          setLoading(false)
-        }
+        trackEvent(tracks.editedUserProfile())
+      } catch (error) {
+        console.error("Failed to update profile", error)
+      } finally {
+        setLoading(false)
+      }
 
-        InteractionManager.runAfterInteractions(() => {
-          onSuccess?.()
-        })
-        navigation.goBack()
-      },
-      validationSchema: editMyProfileSchema,
-    })
+      InteractionManager.runAfterInteractions(() => {
+        onSuccess?.()
+      })
+      navigation.goBack()
+    },
+    validationSchema: editMyProfileSchema,
+  })
+
+  const { handleSubmit, handleChange, dirty, values } = formikBag
 
   // We want to keep the "Save" button enabled as soon as the user edits an input
   const touched = useHasBeenTrue(dirty)
@@ -252,65 +242,9 @@ export const MyProfileEditForm: React.FC<MyProfileEditFormProps> = ({ onSuccess 
           </Flex>
           <Flex m={2}>
             <Join separator={<Spacer y={2} />}>
-              <Input
-                ref={nameInputRef}
-                title="Full name"
-                onChangeText={handleChange("name")}
-                onBlur={() => validateForm()}
-                error={errors.name}
-                returnKeyType="next"
-                value={values.name}
-                onSubmitEditing={() => {
-                  locationInputRef.current?.focus()
-                }}
-              />
-
-              <LocationAutocomplete
-                allowCustomLocation
-                inputRef={locationInputRef}
-                title="Primary location"
-                placeholder="City name"
-                returnKeyType="next"
-                onSubmitEditing={() => {
-                  professionInputRef.current?.focus()
-                }}
-                displayLocation={buildLocationDisplay(values.location)}
-                onChange={({ city, country, postalCode, state, stateCode, coordinates }) => {
-                  setFieldValue("location", {
-                    city: city ?? "",
-                    country: country ?? "",
-                    postalCode: postalCode ?? "",
-                    state: state ?? "",
-                    stateCode: stateCode ?? "",
-                    coordinates,
-                  })
-                }}
-              />
-
-              <Input
-                ref={professionInputRef}
-                title="Profession"
-                onChangeText={handleChange("profession")}
-                onBlur={() => validateForm()}
-                error={errors.name}
-                returnKeyType="next"
-                value={values.profession}
-                placeholder="Profession or job title"
-                onSubmitEditing={() => {
-                  relevantPositionsInputRef.current?.focus()
-                }}
-              />
-
-              <Input
-                ref={relevantPositionsInputRef}
-                title="Other Relevant Positions"
-                onChangeText={handleChange("otherRelevantPositions")}
-                onBlur={() => validateForm()}
-                error={errors.name}
-                returnKeyType="next"
-                value={values.otherRelevantPositions}
-                placeholder="Memberships, institutions, positions"
-              />
+              <FormikProvider value={formikBag}>
+                <UserProfileFields />
+              </FormikProvider>
 
               <ProfileVerifications
                 isIDVerified={!!me?.isIdentityVerified}
